feat: add cli logic to admin and the user cli crates
