+++ crates/zkbtc-admin/Cargo.toml
@@ -1,8 +1,31 @@
 [package]
-name = "zkbtc-admin"
+name = "zkbitcoin-admin-cli"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+anyhow = "1.0.75"
+bitcoin = { version = "0.31.0", features = [
+    "serde",
+    "bitcoinconsensus",
+], git = "https://github.com/mimoo/rust-bitcoin/", branch = "mimoo/fix_0_31" }
+clap = { version = "4.4.10", features = ["derive", "env"] }
+env_logger = "0.10.1"
+log = "0.4.20"
+tempdir = "0.3.7"
+tokio = { version = "1.34", features = [
+    "fs",
+    "rt",
+    "rt-multi-thread",
+    "macros",
+] }
+zkbitcoin-core = { path = "../zkbtc-core"}
+
+
+[patch.crates-io]
+# see docs/serialization.md
+bitcoin = { git = "https://github.com/mimoo/rust-bitcoin/", branch = "mimoo/fix_0_31" }
+# see docs/mpc_and_taproot.md
+frost-core = { git = "https://github.com/mimoo/frost", branch = "mimoo/fix5" }
